+++ platform/example/i2c.c
@@ -5,13 +5,13 @@
 
 #include "platform/i2c.h"
 
-bool i2c_setup(u32 sda, u32 scl, u32 freq) {
+bool i2c_setup(i16 sda, i16 scl, u32 freq) {
     // This function will be called before executing any other I2C-related functions on a given SDA/SCL pair.
     // It should configure the given SDA and SCL pins for I2C communication at the given frequency (in Hz).
     // It should return true if the setup was successful, false if not.
 }
 
-bool i2c_read(u32 sda, u32 scl, byte addr, byte reg, byte dest[], size_t len) {
+bool i2c_read(i16 sda, i16 scl, byte addr, byte reg, byte dest[], size_t len) {
     // This function should read `len` bytes from an 8-bit I2C device address `addr` at an 8-bit I2C register `reg` and
     // store them in the array `dest[]`. For example, if `addr` is 0x69, `reg` is 0x01, and `len` is 1, this function
     // should initiate an I2C transaction to read 1 byte from register 0x01 of device 0x69 and store it in `dest[0]`.
@@ -20,7 +20,7 @@ bool i2c_read(u32 sda, u32 scl, byte addr, byte reg, byte dest[], size_t len) {
     // successful, false if not.
 }
 
-bool i2c_write(u32 sda, u32 scl, byte addr, byte reg, const byte src[], size_t len) {
+bool i2c_write(i16 sda, i16 scl, byte addr, byte reg, const byte src[], size_t len) {
     // This function should write `len` bytes from the array `src[]` to an 8-bit I2C device address `addr` at an 8-bit
     // I2C register `reg`. For example, if `addr` is 0x69, `reg` is 0x01, and `len` is 1, this function should initiate
     // an I2C transaction to write 1 byte to register 0x01 of device 0x69 with the value `src[0]`. This function should
