+++ platform/pico/uart.c
@@ -20,7 +20,7 @@
  * @param rx the pin number of the RX pin
  * @return the UART instance that the pins lie on, or NULL if the pins do not form a valid UART instance
  */
-static inline uart_inst_t *uart_inst_from_pins(u32 tx, u32 rx) {
+static inline uart_inst_t *uart_inst_from_pins(i16 tx, i16 rx) {
     switch (tx) {
         case 0:
         case 12:
@@ -49,7 +49,7 @@ static inline uart_inst_t *uart_inst_from_pins(u32 tx, u32 rx) {
     }
 }
 
-bool uart_setup(u32 tx, u32 rx, u32 baud) {
+bool uart_setup(i16 tx, i16 rx, u32 baud) {
     uart_inst_t *uart = uart_inst_from_pins(tx, rx);
     if (!uart) {
         return false;
@@ -67,7 +67,7 @@ bool uart_setup(u32 tx, u32 rx, u32 baud) {
     return true;
 }
 
-char *uart_read(u32 tx, u32 rx) {
+char *uart_read(i16 tx, i16 rx) {
     uart_inst_t *uart = uart_inst_from_pins(tx, rx);
     if (!uart) {
         return NULL;
@@ -97,7 +97,7 @@ char *uart_read(u32 tx, u32 rx) {
     return buf;
 }
 
-bool uart_write(u32 tx, u32 rx, const char *str) {
+bool uart_write(i16 tx, i16 rx, const char *str) {
     uart_inst_t *uart = uart_inst_from_pins(tx, rx);
     if (!uart) {
         return false;
